logging-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
hazelcast-python-client==5.2